Mount /_synapse/admin/v1/users/{userId}/media admin API on media workers only
